duckduckgo#510 add JustSwipeCallback abstract class
